Simplify finddup documentation
Remove LICENSE section from POD, as it’s the same license as the rest of the
package.

Generate the man page with pod2man: this gets the full documentation,
whereas help2man doesn’t (I suspect the previous man page was manually
improved, or not actually generated with help2man).